refactor: move card preset to seperate file
